Rename files to use a different name of the tool
