+++ kicad_plugin/kiri/kiri.png
