+++ kiri.js
