+++ misc/kiri_pcb.png
