+++ misc/kiri_sch.png
